pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "antigravity-remote"
version = "2.0.0"
description = "Remote control your Antigravity AI assistant via Telegram"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Kubrat" }
]
keywords = ["telegram", "bot", "antigravity", "remote", "ai", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
]

dependencies = [
    "websockets>=12.0",
    "mss>=9.0.0",
    "pyautogui>=0.9.54",
    "pygetwindow>=0.0.9",
    "psutil>=5.9.0",
    "pyperclip>=1.8.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
voice = [
    "SpeechRecognition>=3.10.0",
]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
antigravity-remote = "antigravity_remote.__main__:main"

[project.urls]
Homepage = "https://github.com/kubrat/antigravity-remote"
Repository = "https://github.com/kubrat/antigravity-remote"

[tool.hatch.build.targets.wheel]
packages = ["antigravity_remote"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
